Avoid hammering the server with a tight loop
In handling the `wait` option, we are looping without pause as we check the
pod status. We should pause after each iteration of the loop to avoid undue
resource consumption.

This would be better handled with a kubernetes watch operation, but the
`openshift_client` package explicitly does not support that [1]:

> This API is not intended to implement something as complex as a controller.
> For example, it does not implement watch functionality.

[1]: https://github.com/openshift/openshift-client-python/blob/main/README.md#overview